tidy up utils and fix issue where computer cheated by getting to do special ability AND spell. get rid of possibly_activate_special_ability and just do its stuff inline to prevent dumb circular import workarounds
