+++ game.py
@@ -2,10 +2,12 @@ import json
 import os
 import random
 import time
-from typing import Any, Callable, Optional
+from typing import Optional
 
-from character import Character, SpecialChoice, SpellChoice, SpellDimension
-from game_macros import CHARACTERS_DIR, OPPONENT_SPECIAL_ABILITY_CHANCE
+from character import Character
+from game_macros import (CHARACTERS_DIR, OPPONENT_SPECIAL_ABILITY_CHANCE,
+                         SpecialChoice, SpellChoice, SpellDimension,
+                         confirm_input_choice, did_it_happen, get_input_choice)
 from special_abilities import SpecialAbility
 
 
@@ -22,93 +24,16 @@ class Game:
             character = Character(name=name.title())
             self.all_characters[name] = character
 
-    def get_input_choice(
-        self,
-        prompt: str,
-        choices: dict[str, Any],
-        capitalize_choice: bool = True,
-        offer_random_choice: bool = False,
-    ) -> str:
-        """Given a custom prompt and list of text choices, prompt user to
-        make a choice, and insist that they do so correctly until a proper
-        one can be returned.
-        """
-        input_choices = dict(enumerate(choices))
-        random_choice_index = len(choices)
-        if offer_random_choice:
-            input_choices[random_choice_index] = "Choose for me! 🔮"
-
-        choice = None
-        while choice is None:
-            print(prompt)
-            for idx, item in input_choices.items():
-                print(f"{idx}: {item.title() if capitalize_choice else item}\n")
-
-            choice_input = input(">>> ")
-            try:
-                choice = int(choice_input.strip())
-            except Exception:
-                print("Please choose a number in the given range.")
-                choice = None
-                continue
-
-            # ...but the enum made this 'interface' easy to validate.
-            if choice not in input_choices:
-                print("Please choose a number in the given range.")
-                choice = None
-
-        if choice == random_choice_index and offer_random_choice:
-            del input_choices[random_choice_index]
-            return random.choice(list(input_choices.values()))
-
-        return input_choices[choice]
-
-    def confirm_input_choice(
-        self,
-        choice: str | int,
-        prompt: str,
-        deny_func: Callable,
-        deny_func_kwargs: Optional[dict[str, Any]] = None,
-    ) -> str:
-        """Print prompt presumably associated with some choice
-        (e.g., info about a chosen character), and ask for y/n confirmation.
-
-        Call given deny_func to custom 'reset' if they do not confirm.
-        """
-        deny_func_kwargs = deny_func_kwargs or {}
-        confirmed_choice = None
-
-        while confirmed_choice is None:
-            print(prompt)
-            print(f"Confirm choice? Type y or n.")
-
-            confirm = input(">>> ")
-            try:
-                confirm = confirm.strip().lower()
-            except Exception:
-                print('Please type "y" or "n"')
-                continue
-
-            if confirm == "y":
-                confirmed_choice = choice
-            elif confirm == "n":
-                return deny_func(**deny_func_kwargs)
-            else:
-                print('Please type "y" or "n"')
-                continue
-
-        return confirmed_choice
-
     def select_character(
         self, prompt: str = "Press a key to choose a character: \n"
     ) -> str:
-        chosen_input = self.get_input_choice(
+        chosen_input = get_input_choice(
             prompt=prompt,
             choices=self.all_characters,
             capitalize_choice=True,
             offer_random_choice=True,
         )
-        chosen_confirmed = self.confirm_input_choice(
+        chosen_confirmed = confirm_input_choice(
             choice=chosen_input,
             prompt=f"{self.all_characters[chosen_input].ascii_art}\n\n{self.all_characters[chosen_input].bio}\n",
             deny_func=self.select_character,
@@ -154,8 +79,8 @@ class Game:
 
     def player_turn(self) -> None:
         spell_infos = self._construct_player_spell_choices()
-        spell = self.get_input_choice(
-            prompt="Choose your spell:\n", choices=spell_infos, capitalize_choice=False
+        spell = get_input_choice(
+            prompt="\nChoose your spell:\n", choices=spell_infos, capitalize_choice=False
         )
         choice = spell_infos[spell]
 
@@ -164,9 +89,9 @@ class Game:
             self.hit(self.opponent, dimension, max_hit)
             self.opponent.possibly_react()
 
-        if isinstance(choice, SpecialChoice):
+        elif isinstance(choice, SpecialChoice):
             description, effect = choice.description, choice.effect
-            self.confirm_input_choice(
+            confirm_input_choice(
                 choice=spell,
                 prompt=description,
                 deny_func=self.player_turn,
@@ -179,30 +104,43 @@ class Game:
     def opponent_turn(self) -> None:
         self.opponent.possibly_taunt()
 
-        # the opponent of the opponent is of course the player, so keep that
-        # in mind when returning player, opponent result format
-        (
-            modified_opponent_as_player,
-            modified_player_as_opponent,
-        ) = self.opponent.possibly_activate_special_ability(
-            chance=OPPONENT_SPECIAL_ABILITY_CHANCE,
-            human_opponent=self.player,
-        )
-        self.opponent, self.player = (
-            modified_opponent_as_player,
-            modified_player_as_opponent,
-        )
+        if self.opponent.special_abilities_info and did_it_happen(
+            OPPONENT_SPECIAL_ABILITY_CHANCE
+        ):
+            special_ability = SpecialAbility(
+                player=self.opponent,
+                opponent=self.player,
+                effect=random.choice(
+                    [
+                        info["effect"]
+                        for info in self.opponent.special_abilities_info.values()
+                    ]
+                ),
+            )
+            # The opponent of the opponent is of course the player, let's be super
+            # clear about player, opponent result format returned by perform()
+            (
+                modified_opponent_as_player,
+                modified_player_as_opponent,
+            ) = special_ability.perform(is_computer=True)
+            self.opponent, self.player = (
+                modified_opponent_as_player,
+                modified_player_as_opponent,
+            )
+        # If the opponent does not activate a special ability, they just choose a spell:
+        else:
+            spell_info = self.opponent.magic_info["deals"]
+            # Recall that not everyone can deal every kind, as a cost to being
+            # super strong in some.
+            able_dimensions = [
+                dim for dim, info in spell_info.items() if info["spells"]
+            ]
+            dimension = random.choice(able_dimensions)
+            spell = random.choice(spell_info[dimension]["spells"])
 
-        spell_info = self.opponent.magic_info["deals"]
-        # Recall that not everyone can deal every kind, as a cost to being
-        # super strong in some.
-        able_dimensions = [dim for dim, info in spell_info.items() if info["spells"]]
-        dimension = random.choice(able_dimensions)
-        spell = random.choice(spell_info[dimension]["spells"])
-
-        print(f'{self.opponent.name} chooses: "{spell}"')
-        time.sleep(1)
-        self.hit(self.player, dimension, max_hit=spell_info[dimension]["amount"])
+            print(f'\n{self.opponent.name} chooses: "{spell}"')
+            time.sleep(1)
+            self.hit(self.player, dimension, max_hit=spell_info[dimension]["amount"])
 
     def play(self) -> None:
         player_choice = self.select_character()
